remove pod informer
